+++ src/ts/Adders/BreadcrumbsArrowsAdder.ts
@@ -39,17 +39,16 @@ export default class BreadcrumbsArrowsAdder implements Adder {
     }
 
     private getTitle(elem: HTMLElement): string {
-        return elem.innerText
-            .replace('« ', '')
-            .replace(' »', '')
-            .trim()
+        return elem.innerText.replace('« ', '').replace(' »', '').trim()
     }
 
     private getLink(elem: HTMLElement): string {
         const anchor = elem.querySelector<HTMLAnchorElement>('a')
 
         if (!anchor) {
-            console.warn('[PHP Revival]: BreadcrumbsArrowsAdder: anchor not found in breadcrumbs arrow link')
+            console.warn(
+                '[PHP Revival]: BreadcrumbsArrowsAdder: anchor not found in breadcrumbs arrow link',
+            )
             return '#'
         }
 
@@ -57,7 +56,11 @@ export default class BreadcrumbsArrowsAdder implements Adder {
     }
 
     private insertNewArrows(): void {
-        const rightLink = this.createArrowLink(this.nextTitle, this.nextLink, 'right')
+        const rightLink = this.createArrowLink(
+            this.nextTitle,
+            this.nextLink,
+            'right',
+        )
         const leftLink = this.createArrowLink(this.prevTitle, this.prevLink, 'left')
         const container = this.createLinksContainer()
 
@@ -74,7 +77,11 @@ export default class BreadcrumbsArrowsAdder implements Adder {
         return container
     }
 
-    private createArrowLink(title: string, link: string, direction: 'left' | 'right'): HTMLElement {
+    private createArrowLink(
+        title: string,
+        link: string,
+        direction: 'left' | 'right',
+    ): HTMLElement {
         const a = document.createElement('a')
 
         a.href = link