+++ src/ts/Adders/CopyButtonAdder.ts
@@ -8,7 +8,9 @@ const REMOVE_TOOLTIP_AFTER = 1000
 
 export default class CopyButtonAdder implements AdderInterface {
     public add(): void {
-        const targetsList = document.querySelectorAll<HTMLElement>(conf.selectors.targetForCodeExamples)
+        const targetsList = document.querySelectorAll<HTMLElement>(
+            conf.selectors.targetForCodeExamples,
+        )
         const targets = Array.from(targetsList)
 
         if (!targets.length) {
@@ -25,7 +27,9 @@ export default class CopyButtonAdder implements AdderInterface {
 
     private listenForButtonClick(targets: HTMLElement[]): void {
         for (const target of targets) {
-            const copyIcon = target.querySelector<HTMLElement>(conf.selectors.copyCodeIcons)
+            const copyIcon = target.querySelector<HTMLElement>(
+                conf.selectors.copyCodeIcons,
+            )
 
             if (!copyIcon) {
                 continue
@@ -45,7 +49,11 @@ export default class CopyButtonAdder implements AdderInterface {
             })
     }
 
-    private showTooltip(target: HTMLElement, text: string, isSuccess: boolean): void {
+    private showTooltip(
+        target: HTMLElement,
+        text: string,
+        isSuccess: boolean,
+    ): void {
         const tooltip = this.createTooltipElement(text, isSuccess)
 
         target.appendChild(tooltip)