+++ src/ts/templates/homeLinkTemplate.ts
@@ -28,13 +28,13 @@ function createLink(img: Element, span: Element, link: HomeLink): HTMLDivElement
 
     a.href = link.href
     a.target = '_blank'
-    a.className = 'php-revival-home-links-section__link'
 
     a.appendChild(img)
     a.appendChild(span)
 
     const div = document.createElement('div')
 
+    div.className = 'php-revival-home-links-section__link'
     div.dataset.url = link.href
     div.appendChild(a)