+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## v2.9.0 (2024-09-20)
+- Formatted TypeScript code with Prettier
 
 ## v2.8.5 (2024-09-10)
 - Fixed styles for the "Changelog" section on function page. `table-layout: fixed` wasn't a good solution for that table