+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Release Notes
 
-## v2.9.0 (2024-09-20)
+## v2.9.0 (2024-09-22)
 - Formatted TypeScript code with Prettier
 - Improved home sidebar by splitting links into 3 sections: "Events", "Recommended" and "Social"
+- Add showing nested links in the home sidebar where it's needed
 
 ## v2.8.5 (2024-09-10)
 - Fixed styles for the "Changelog" section on function page. `table-layout: fixed` wasn't a good solution for that table